+++ ChangeLog.md
@@ -9,6 +9,7 @@
   ([#59](https://github.com/davep/hike/pull/59))  
 - Added <kbd>ctrl</kbd>+<kbd>r</kbd> as an alternative binding for jumping
   to the document. ([#60](https://github.com/davep/hike/pull/60))
+- Added support for scrolling the markdown up/down half a page.
 
 ## v0.7.0